Revamp indexing
We're now storing the archived frameworks and resulting DB in blob storage.
Also, the indexer can now reason about which shared frameworks belong to which
TFM. What's missing is exposing this information in the DB. The best way to do
this is likely to add some data to the assembly that stays how the assembly must
be referenced.