Speed up package indexing
Instead of pulling down all package ids and version from nuget.org and then
filtering it down to just the ones owned by us, start with the package ids
from the owner mapping, filter down to the package ids owned by us and then
ask for the latest version.

This also ensures we're indexing the latest stable and the latest preview, if
the preview is newer.