Incrementally write packages
Instead of bulk inserting all packages following by its usages, this will
process the packages and their usages together. This allows us to introduce a
timeout for crawling such that we have enough time left to upload the DB.

This way, if one day crawling exceeds the GitHub Action limitation of six
hours, we were at least making progress so that we eventually complete
indexing.

This will also upload the DB before exporting the results. This allows avoids
having to use a transaction that we need to undo, which exposes a bug in
SQLite and ends up corrupting the DB when rolling back.