Revamping usage analysis
We regularly run usage analysis against NuGet packages. Currently it only covers
APIs. Sometimes we have the need for custom IL analysis.

In order to fit this into the data model, we define it as a feature collector
where each feature is reported as a GUID. For simple API usage, the GUID simply
refers to the API.

For non-API feature detection, we could just define an arbitrary GUID. For
example, when measuring if any multi-dimensional arrays are used, we could
define a fixed GUID for that.

For features that need to report multiple different values, we could define it
as a hash-derived GUID from suitable inputs. For example, let's say we wanted
to measure the use of types and number of dimensions in arrays, we could define
it as a hash of the fixed GUID for the multidimensional array feature, the type
GUID, and the number of dimensions.

So long we can statically compute a hash to look up, this scheme will work.

This requires a few modifications:

1. The data no longer refer to the GUIDs as APIs but as features.

2. We also abstracted out the notion of packages -- we already have duplication
   with the way we collect the API telemetry from .NET Upgrade Planner. We now
   refer to this as "reference units". The usage percentage is simply the
   number of reference units using a given feature divided by the total number
   of reference units.

3. We now also handle versioning of the collector. When we add a new feature, we
   don't want to have to re-index the world in order to have data consistency.
   Equally, we don't want to report faulty percentages by considering
   previously indexed reference units that didn't consider this new feature. To
   solve this, reference units store the version of the collector and features
   store the version of the collector that introduced them. When computing a
   percentage, we only consider reference units for a given feature if its
   version is equal to or higher than the feature version.

Missing:

1. We need to change GenUsagePlanner to use the new UsageDatabase.

2. We need a way to embed this more generalized usage data into API Catalog.
   This likely requires schema changes.