Rename VersionIntroduced to VersionRequired
If we modify a collector and want to require re-indexing, we want to
just bump its version. So it makes sense to have the mechanism be that
a collector can just require a minimum version.

Equally, at that point we want to know what the current version so
instead of having the collector set compute the max we want to just
expose a fixed number there.

Of course, we still want the rule that the collector set version is the
max version required of all collectors, but we can enforce that via unit
test.